apps/api/requirements.txt
fastapi==0.115.4
uvicorn[standard]==0.30.6
sqlalchemy==2.0.36
sqlmodel==0.0.22
asyncpg==0.30.0
alembic==1.14.0
pydantic==2.9.2
pydantic-settings==2.6.1
python-jose[cryptography]==3.3.0
cryptography==43.0.3
redis==5.2.0
boto3==1.35.54
qrcode[pil]==7.4.2
python-multipart==0.0.12
